requirements.txt
boto==2.49.0
boto3==1.9.86
botocore==1.12.86
bz2file==0.98
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cld2-cffi==0.1.4
cymem==2.0.2
cytoolz==0.9.0.1
de-core-news-sm==2.0.0
decorator==4.3.2
dill==0.2.9
Distance==0.1.3
docutils==0.14
gensim==3.7.1
idna==2.8
jmespath==0.9.3
kafka-python==1.4.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multi-rake==0.0.1
murmurhash==1.0.1
networkx==2.2
nltk==3.4
numpy==1.16.1
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
pycparser==2.19
pydeepl==0.13
pymongo==3.7.2
pyrsistent==0.14.9
pysummarize==0.6.0
python-dateutil==2.7.5
regex==2018.1.10
requests==2.21.0
s3transfer==0.1.13
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
spacy==2.0.18
summa==1.2.0
textblob==0.15.2
textblob-de==0.4.3
thinc==6.12.1
toolz==0.9.0
tqdm==4.30.0
ujson==1.35
urllib3==1.24.1
wrapt==1.10.11
